small fix cuz fuck windows

+++ database/migrations/2020_03_06_150929_create_payment_options.php
@@ -14,7 +14,7 @@ class CreatePaymentOptions extends Migration
     public function up()
     {
         Schema::create('payment_options', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->integer('id');
             $table->string('title');
             $table->string('price');
             $table->longText('description');
